% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{create_cv_folds_old}
\alias{create_cv_folds_old}
\title{Code adapted from HPprediction package}
\usage{
create_cv_folds_old(
  Z,
  n = 10,
  min_per_col = 2,
  min_per_row = 2,
  get_zeros = FALSE
)
}
\description{
Code adapted from HPprediction package
}
