% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Data \code{dat}}
\format{
\subsection{\code{dat}}{

A \code{matrix} with 100 rows and 14 columns:
\describe{
\item{Rows}{Each row represents a different species.}
\item{Columns}{Each column represents a different site.}
\item{Entries}{Binary values where 1 indicates interaction/link and 0 indicates no interaction/no link between node pairs.}
}
}
}
\source{
Data is of own creation.
}
\usage{
dat
}
\description{
A binary bipartite network matrix representing interactions between different types of nodes.
}
\keyword{datasets}
