% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2.R
\docType{data}
\name{dat2}
\alias{dat2}
\title{Data 2 \code{dat2}}
\format{
\subsection{\code{dat2}}{

A \code{matrix} with 84 rows and 145 columns:
\describe{
\item{Rows}{Each row represents a different species of the A type.}
\item{Columns}{Each column represents a different species of the B type.}
\item{Entries}{Binary values where 1 indicates interaction/link and 0 indicates no interaction/no link between node pairs.}
}
}
}
\source{
Data 2 is of own creation.
}
\usage{
dat2
}
\description{
An artificial binary bipartite network matrix representing interactions between different types of nodes.
}
\keyword{datasets}
