% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsbm_results.R
\name{get_hsbm_results}
\alias{get_hsbm_results}
\title{Extract and summarize results from HSBM predictions.}
\usage{
get_hsbm_results(hsbm_output, input_names = TRUE, na_treatment = "na_to_0")
}
\arguments{
\item{hsbm_output}{An object of class \code{hsbm.predict} containing the predictions made by the HSBM model.}

\item{input_names}{(\emph{optional, default} \code{TRUE}) \cr
A \code{logical} value indicating whether to include the names of the nodes in the output. If \code{TRUE}, node names are extracted from the original data and added to the results.}

\item{na_treatment}{(\emph{optional, default} \code{"na_to_0"}) \cr
A \code{character} string specifying how to handle \code{NA} values in links derived from HSBM predictions. Options include \code{"na_to_0"}, which interprets \code{NA} values as no evidence of a existing link and assigns them a value of zero; \code{"keep_na"}, which retains \code{NA} if there is at least one \code{NA} among the predictions of folds, maintaining uncertainty; and \code{"ignore_na"}, which computes the average of predictions while ignoring \code{NA} values. This parameter provides flexibility in managing \code{NA} values within the final (averaged) predicted network.}
}
\value{
An object of class \code{hsbm.predict} with additional elements:
\itemize{
\item \code{$predictions$res_folds} A \code{list} where each element is a \code{data.frame} containing the results of the predictions for each fold, including predicted probabilities and edge types.  						#@@JMB No recuerdo si qué diferencia había con $predictions$probs. Parece idéntico
\item \code{$predictions$res_averaged} A \code{data.frame} summarizing the averaged results across all folds.
\itemize{
\item \code{v1} The index of the first type of node.
\item \code{v2} The index of the second type of node.
\item \code{v1_names} Names of the first type of nodes.
\item \code{v2_names} Names of the first type of nodes.
\item \code{p} The average predicted probabilities for each edge.
\item \code{sd} The standard deviation of predictions for each edge.
\item \code{range} The range of predicted probabilities for each edge.
\item \code{edge_type} Assigned type of edge after edge prediction: 'reconstructed' for undocumented edge (0s), and 'documented' for observed links (1s).
\item \code{nr_na} The number of folds in which each edge was \code{NA} (i.e., no prediction was made).
}
}
}
\description{
This function processes the output from an HSBM prediction to generate and summarize the results. It combines predictions across all cross-validation folds, computes the average probability, standard deviation, range of probabilities, and edge_type for each edge/link across the folds.
}
\examples{
# Load example HSBM prediction results
data(myPred, package = "sabinaHSBM")

# Extract and summarize the results
myResults <- get_hsbm_results(hsbm_output = myPred)

# View the final averaged matrix
averaged_matrix <- myResults$predictions$res_averaged

}
\seealso{
\code{\link{hsbm.predict}}
}
