% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsbm_results.R
\name{hsbm.reconstructed}
\alias{hsbm.reconstructed}
\title{Generate a reconstructed bipartite binary matrix from Hierarchical Stochastic Block Model (HSBM)}
\usage{
hsbm.reconstructed(
  hsbm_out,
  pred_all = FALSE,
  rm_documented = FALSE,
  threshold = "prc_closest_topright",
  new_matrix_method = "average_thresholded",
  custom_threshold = NULL
)
}
\arguments{
\item{hsbm_out}{An object of class \code{hsbm.output} containing the output from the HSBM analysis, including predictions and input data.}

\item{pred_all}{(\emph{optional, default} \code{FALSE}) \cr
A \code{logical} value indicating whether to use all predictions for reconstruction.    #@@JMB terminar de aclarar}

\item{rm_documented}{(\emph{optional, default} \code{FALSE}) \cr
A \code{logical} value indicating whether to remove documented entries (1s).    #@@JMB terminar de aclarar}

\item{threshold}{(\emph{optional, default} \code{"prc_closest_topright"}) \cr
A \code{character} string or \code{numeric} value specifying the method to determine the threshold for binary classification of predictions. See Details for available options.}

\item{new_matrix_method}{(\emph{optional, default} \code{"average_thresholded"}) \cr
A \code{character} string specifying the method for creating the new reconstructed matrix. Options include \code{"average_thresholded"} and \code{"ensemble_binary"}.}

\item{custom_threshold}{(\emph{optional, default} \code{NULL}) \cr
A \code{numeric} value specifying a custom threshold for binary classification. If \code{NULL}, the default threshold is used used based on the \code{new_matrix_method}.    #@@JMB igual podemos llamar a esto ensemble_binary_threshold para que quede más claro. Yo no le entiendo muy bien así que revisa que esté la explicación correcta}
}
\value{
An object of class \code{hsbm.reconstructed} containing the reconstructed matrices and related statistics:
\itemize{
\item \code{$data} A \code{matrix} of the original input data.
\item \code{$pred_mats} A \code{list} of \code{matrix}, each containing predicted probabilities for edges/links from different cross-validation folds.
\item \code{$reconstructed_bin_mats} A \code{list} of binary reconstructed matrices (0s or 1s) for each fold, based on the applied threshold.
\item \code{$reconstructed_stats} A list of reconstruction statistics for each fold.
\item \code{$tb} A \code{data.frame} with combined statistics for all folds. It includes the following columns:
\itemize{
\item \code{folds}: The index of the cross-validation fold.
\item \code{auc}: The Area Under the Curve (AUC) of the Receiver Operating Characteristic (ROC) curve for the fold.
\item \code{aucpr}: The Area Under the Curve (AUC) of the Precision-Recall Curve (PRC) for the fold.
\item \code{yPRC}: The baseline of Precision-Recall Curve for the fold.
\item \code{thresh}: The binary classification threshold applied to the predicted probabilities to convert them into binary classifications (0 or 1) for the fold.
\item \code{n_heldout}: The number of edges/links in the held-out set of the cross-validation fold, i.e., the edges that were excluded from the model training and used for evaluation.
\item \code{pred_held_ones}: The proportion of held-out edges/links that were correctly predicted as 1s by the model.
\item \code{n_ones}: The total number of positive edges/links in the original data.
\item \code{pred_tot_ones}: The proportion of positive edges/links in the original data that were correctly predicted as positive by the model.
\item \code{total_pred_ones}: The total number of positive edges/links predicted by the model in the reconstructed matrix.  		#@@JMB estqo es así?
\item \code{precision}: The precision of the model, calculated as the ratio of true positives to the sum of true positives and false positives.
\item \code{sens}: The sensitivity (or recall) of the model.
\item \code{spec}: The specificity of the model.
\item \code{ACC}: The overall accuracy of the model.
\item \code{ERR}: The error rate of the model.
\item \code{tss}: The True Skill Statistic (TSS).
}
\item \code{$new_mat} A \code{matrix} representing the final reconstructed binary matrix using the specified \code{new_matrix_method} and \code{threshold}.
\item \code{$threshold} A \code{character} string indicating the method used to determine the binary classification threshold.
}
}
\description{
This function generates a reconstructed bipartite binary matrix based on the HSBM model.
}
\details{
\itemize{
\item The \code{pred_all} parameter indicates whether to use all predictions or a subset for the reconstruction.
\item The \code{rm_documented} parameter controls whether documented entries (1s) are removed from the reconstruction.  #@@JMB terminar de aclarar
\item The \code{threshold} parameter defines the method for determining the binary classification threshold. Valid options are:
\itemize{
\item \code{"roc_youden"}: Maximizes the Youden's J statistic (sensitivity + specificity - 1).
\item \code{"roc_closest_topleft"}: Minimizes the distance to the top-left corner in the ROC space.
\item \code{"roc_equal_sens_spec"}: Equalizes sensitivity and specificity.
\item \code{"roc_no_omission"}: Maximizes the threshold with no false negatives.
\item \code{"prc_min_rec_prec"}: Minimizes the sum of recall and precision.
\item \code{"prc_equal_rec_prec"}: Equalizes recall and precision.
\item \code{"prc_closest_topright"}: Minimizes the distance to the top-right corner in the PRC space.
\item \code{"prc_max_F1"}: Maximizes the F1 score.
\item Alternatively, a numeric value between 0 and 1 can be provided as a custom threshold.
}
\item The \code{new_matrix_method} parameter specifies the method used to generate the new reconstructed matrix. Valid options are:
\itemize{
\item \code{"average_thresholded"}: This method averages the predicted matrices and then applies a threshold to create a binary matrix. The threshold can be set using the \code{custom_threshold} parameter or defaults to the average threshold calculated from the folds.
\item \code{"ensemble_binary"}: This method creates a binary matrix by taking an ensemble approach, where each entry is set to 1 if it exceeds the specified threshold in any of the prediction matrices. The threshold can be set using the \code{custom_threshold} parameter or defaults to a predefined value (e.g., 0.1).
}
}
}
\examples{
# Load example HSBM prediction results
data(myPred, package = "sabinaHSBM")

myReconst <- hsbm.reconstructed(myPred)

# View the final reconstructed binary matrix
reconstructed_matrix <- myReconst$new_mat

}
\seealso{
\code{\link{hsbm.predict}}
}
