% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_matrix.R
\name{plot.hsbm.input}
\alias{plot.hsbm.input}
\title{Plot binary bipartite matrix from \code{hsbm.input} objects}
\usage{
\method{plot}{hsbm.input}(hsbm_input, col = "red4", order_mat = TRUE, ...)
}
\description{
Use \code{plot_interaction_matrix} to plots a binary bipartite
matrix from a \code{hsbm.input} object.
}
\seealso{
\code{\link{plot_interaction_matrix}}
}
