% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hsbm.predict}
\alias{summary.hsbm.predict}
\title{Summary for hsbm.predict Objects}
\usage{
\method{summary}{hsbm.predict}(object, ...)
}
\arguments{
\item{object}{An object of class \code{hsbm.predict}.}

\item{...}{Additional arguments (not used).}
}
\value{
A data frame containing:
\describe{
\item{n_rows}{Number of rows (first set of nodes).}
\item{n_cols}{Number of columns (second set of nodes).}
\item{n_links}{Number of observed links in the binary input matrix.}
\item{rows_single_link}{Number of rows with only one link.}
\item{cols_single_link}{Number of columns with only one link.}
\item{possible_links}{Total number of possible links in the input matrix.}
}
}
\description{
Provides a summary for objects of class \code{hsbm.predict}.
}
\seealso{
\code{\link{hsbm.predict}}
}
