% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hsbm.reconstructed}
\alias{summary.hsbm.reconstructed}
\title{Summary for hsbm.reconstructed Objects}
\usage{
\method{summary}{hsbm.reconstructed}(object, ...)
}
\arguments{
\item{object}{An object of class \code{hsbm.reconstructed}.}

\item{...}{Additional arguments (not used).}
}
\value{
A data frame containing the following summary statistics:
\describe{
\item{obs_links}{Number of observed links in the original input matrix.}
\item{unobs_links}{Number of unobserved links in the original input matrix.}
\item{pred_links}{Number of predicted links in the reconstructed binary matrix.}
\item{kept_links}{Number of links retained (unchanged) in the reconstructed matrix.}
\item{spurious_links}{Number of links removed (false positives) in the reconstructed matrix.}
\item{missing_links}{Number of new links added (false negatives) in the reconstructed matrix.}
\item{mean_RLF}{Mean Recovery Link Fraction (proportion of held-out links correctly predicted as missing links) across all folds.}
\item{mean_auc}{Mean AUC across all folds.}
\item{mean_aucpr}{Mean AUC-PR across all folds.}
\item{mean_yPRC}{Mean Precision-Recall Curve baseline across all folds.}
\item{mean_prec}{Mean precision across all folds.}
\item{mean_sens}{Mean sensitivity across all folds.}
\item{mean_spec}{Mean specificity across all folds.}
\item{mean_ACC}{Mean accuracy across all folds.}
\item{mean_ERR}{Mean error rate across all folds.}
\item{mean_tss}{Mean True Skill Statistic (TSS) across all folds.}
}
}
\description{
Provides a summary for objects of class \code{hsbm.reconstructed}.
}
\seealso{
\code{\link{hsbm.reconstructed}}
}
