% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{create_cv_folds}
\alias{create_cv_folds}
\title{Code from HPprediction package}
\usage{
create_cv_folds(
  Z,
  n = 10,
  min.per.col = 1,
  missing.pattern = c("random", "prop.to.col.sums")
)
}
\description{
Code from HPprediction package
}
