% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPred.R
\docType{data}
\name{myPred}
\alias{myPred}
\title{Example HSBM prediction object}
\format{
A \code{list} with the following components:
\describe{
\item{data}{A \code{matrix} representing the binary bipartite input data.}
\item{folds}{A \code{matrix} specifying cross-validation fold assignments for held-out edges/links.}
\item{is_bipartite}{A code{logical} indicator for bipartite/unipartite matrices.}
\item{method}{A \code{character} string indicating the method used for the HSBM analysis, in this case, \code{"conditional_missing"}.}
\item{iter}{An \code{integer} specifying the number of iterations used for the HSBM analysis.}
\item{wait}{A \code{numeric} value specifying the number of iterations used for MCMC equilibration.}
\item{min_dl}{A \code{list} of minimum description length values for each fold.}
\item{probs}{Predicted probabilities for edges/links, stored as \code{data.frame}s, one for each fold.}
\item{groups}{Group assignments for nodes across hierarchical levels, stored as \code{data.frame}s.}
}
}
\source{
Generated using the \code{hsbm.predict()} function in the \code{sabinaHSBM} package.
}
\usage{
data(myPred)
}
\description{
An example of an object of class \code{hsbm.predict}, precomputed using the \code{hsbm.predict()} function. This object is included to facilitate the demonstration of functionality in examples and vignettes without requiring computational overhead.
}
\keyword{datasets}
