% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_matrix.R
\name{plot_interaction_matrix}
\alias{plot_interaction_matrix}
\title{Plot binary bipartite Matrix}
\usage{
plot_interaction_matrix(
  adj_mat,
  order_mat = TRUE,
  y_title = "Rows",
  x_title = "Columns",
  ...
)
}
\arguments{
\item{adj_mat}{A \code{matrix} or \code{data.frame} representing the interaction data to be plotted.}

\item{order_mat}{(\emph{optional, default} \code{TRUE}) \cr
A \code{logical} value indicating whether the matrix should be order by row frequencies.}

\item{y_title}{(\emph{optional, default} \code{"Rows"}) \cr
A \code{character} string specifying the title for the y-axis.}

\item{x_title}{(\emph{optional, default} \code{"Columns"}) \cr
A \code{character} string specifying the title for the x-axis.}

\item{...}{Further arguments for \code{image} R native function}
}
\value{
A plot of the interaction matrix.
}
\description{
This function plots a (recontructed) binary bipartite matrix with options to reorder the matrix and customize the colors and axis titles.
}
\details{
The \code{order_mat} argument indicates whether the interaction matrix should be iteratively reordered to enhance the grouping of links for improved visualization.
}
\examples{
# Load example HSBM prediction results
data(dat, package = "sabinaHSBM")

# Plot the Known matrix
plot_interaction_matrix(dat)

# Plot the Reconstructed matrix
# data(myReconst, package = "sabinaHSBM")
# plot_interaction_matrix(myReconst$new_mat, order_mat = FALSE)

}
