% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsbm_results.R
\name{top_links}
\alias{top_links}
\title{Extract Top Predicted Links}
\usage{
top_links(hsbm_reconstructed, n = 10)
}
\arguments{
\item{hsbm_reconstructed}{An object of class \code{hsbm.reconstructed} containing the results of the HSBM model reconstruction.}

\item{n}{(\emph{optional, default} \code{10}) \cr
An \code{integer} specifying the number of top links to extract.}
}
\value{
A \code{data.frame} with the top \code{n} predicted links (links with the highest probabilities), including the node names, predicted probabilities (\code{p}), and their standard deviations (\code{sd}) as a uncertainty measure.
}
\description{
Extracts the top predicted links from the HSBM model output, allowing for prioritization of likely true links.
}
\examples{
# Load example HSBM prediction results
# Assuming `hsbm_reconstructed` is an object of class `hsbm.reconstructed`
data(myReconst, package = "sabinaHSBM")

top_10_links <- top_links(myPred, n = 10)
print(top_10_links)

}
\seealso{
\code{\link{hsbm.predict}}, \code{\link{hsbm.reconstructed}}
}
